Add files via upload
config.html:
I added a "back" button to the config.html which now leads back to the index.html. 

Esp_radio.ino:
This is an updated version. I didn't manage to fork from the source, so I just renamed and uploaded the new file. Please check and feel free to integrate the code into your codebase.

New Features for TFT-Version with one button are:
-    Long button press (between 300 ms and 2 secs) go back one channel
-    Wrap channel up and down (so back form preset zero goes to last preset)
-    Long press (>2 secs) mutes, every press afterwards unmutes the sound.