Remove ArduinoLog dependency  (#13)
To get rid of ArduinoLog 3rd party dependency. 
To make use of ESP Arduino framework built in logger.
Depending on the platform it's the one which comes from 
Arduino core for ESP32 or Arduino core for ESP8266. 
Documentation updated accordingly.

The original plan to provide ESP-IDF framework compatibility is not yet done, 
as the library implementation uses much more Arduino core for ESP dependent stuff.