Fix load user code bug (#82)
A fix for bug in `loadUserCode`. This seems like incompatible change, as the second argument was introduced, however the previous version is buggy and it's better to get a compilation error, and adjust a code to get rid of bug (if a new loadUserCode was already used). Please also see a discussion here: https://github.com/baldram/ESP_VS1053_Library/discussions/66#discussioncomment-1306049